Moved dust sensor connection to PCB edge.
Removed pullup/down resistors for PWM and tach signals as they are not needed.

Added P-FET + N-FET power control to switch the 12V to the PWM fans as some fans do not switch off when PWM off.

Moved 12V LED to switched fan power line and to underneath the "On" lettering.